+++ src/Scripts/src/client/weavy.js
@@ -888,19 +888,97 @@
          **/
         function frameStatusCheck() {
             var statusUrl = "/client/ping";
+
+            var storageAccessAvailable = 'hasStorageAccess' in document;
+
+            var whenFrameCookiesChecked = new WeavyPromise();
+            var whenFrameCookiesEnabled = new WeavyPromise();
+            var whenFrameReady = new WeavyPromise();
+
             var whenFrameStatusCheck = new WeavyPromise();
 
+            var alertCookie, alertStorage;
+
             if (!weavy.nodes.statusFrame) {
+                weavy.log("Frame Check: Started...", weavy.getId());
                 // frame status checking
-                weavy.nodes.statusFrame = document.createElement("iframe");
-                weavy.nodes.statusFrame.className = "weavy-status-check weavy-hidden";
-                weavy.nodes.statusFrame.style.display = "none";
-                weavy.nodes.statusFrame.id = weavy.getId("weavy-status-check");
-                weavy.nodes.statusFrame.setAttribute("name", weavy.getId("weavy-status-check"));
+                var statusFrame = weavy.nodes.statusFrame = document.createElement("iframe");
+                statusFrame.className = "weavy-status-check weavy-hidden";
+                statusFrame.style.display = "none";
+                statusFrame.id = weavy.getId("weavy-status-check");
+                statusFrame.setAttribute("name", weavy.getId("weavy-status-check"));
+
+                var requestStorageAccess = function () {
+                    var msg = utils.asElement('<span>Third party cookies are required to use this page. </span>')
+                    var msgButton = utils.asElement('<button class="btn" style="pointer-events: auto;">Enable cookies</button>');
+                    var storageAccessWindow;
+
+                    msgButton.onclick = function () {
+                        weavy.log('Frame Check: Opening storage access request');
+                        storageAccessWindow = window.open(weavy.options.url + '/cookie-access', weavy.getId("weavy-storage-access"));
+                        wvy.postal.registerContentWindow(storageAccessWindow, weavy.getId("weavy-storage-access"), weavy.getId());
+                    };
+                    msg.appendChild(msgButton);
+
+                    alertStorage = weavy.alert(msg, true);
+
+                    weavy.one(wvy.postal, "storage-access-granted", { weavyId: true, domain: new URL(weavy.options.url).origin }, function () {
+                        weavy.log("Frame Check: Storage access was granted, authenticating and reloading status check.", alertCookie, alertStorage);
+
+                        if (alertCookie) {
+                            alertCookie.remove();
+                            alertCookie = null;
+                        }
+
+                        if (alertStorage) {
+                            alertStorage.remove();
+                            alertStorage = null;
+                        }
+                        let weavyId = weavy.getId();
+
+                        weavy.authentication.signIn().then(function () {
+                            weavy.debug("Frame Check: reloading status check")
+                            wvy.postal.postToFrame(weavy.getId("weavy-status-check"), weavyId, { name: "reload" });
+                        });
+                    })
+
+                };
+
+                weavy.on(wvy.postal, "user-status", { weavyId: weavy.getId(), windowName: weavy.getId("weavy-status-check") }, function (e, userStatus) {
+                    var cookieIsValid = parseInt(userStatus.id) === parseInt(weavy.authentication.user().id);
+                    weavy.debug("Frame Check: user-status received", cookieIsValid);
+                    whenFrameCookiesChecked.resolve(cookieIsValid);
+
+                    if (!cookieIsValid) {
+                        if (storageAccessAvailable) {
+                            requestStorageAccess();
+                        } else if (!storageAccessAvailable) {
+                            alertCookie = weavy.alert('Allow third party cookies to use this page.');
+                        }
+                    } else {
+                        whenFrameCookiesEnabled.resolve();
+                    }
+                });
+
 
                 weavy.one(wvy.postal, "ready", { weavyId: weavy.getId(), windowName: weavy.getId("weavy-status-check") }, function () {
-                    weavy.log("Frame status check", "√")
+                    weavy.debug("Frame Check: frame ready")
+                    whenFrameReady.resolve();
+                });
+
+                Promise.all([whenFrameReady(), whenFrameCookiesEnabled()]).then(function () {
+                    weavy.log("Frame Check:", "OK");
                     weavy.isBlocked = false;
+
+                    if (alertCookie) {
+                        alertCookie.remove();
+                        alertCookie = null;
+                    }
+
+                    if (alertStorage) {
+                        alertStorage.remove();
+                        alertStorage = null;
+                    }
 
                     /**
                      * Triggered when the frame check is done.
@@ -916,6 +994,7 @@
                 });
 
                 weavy.nodes.container.appendChild(weavy.nodes.statusFrame);
+
                 weavy.whenTimeout(1).then(function () {
                     weavy.nodes.statusFrame.src = weavy.httpsUrl(statusUrl, weavy.options.url);
                     weavy.isBlocked = true;